requirements.txt
pyserial==3.4
PyUserInput==0.1.11
pywin32==227
six==1.14.0
